Refactor task dashboard and introduce MyTasks component
- Removed redundant code from the TaskDashboard component and replaced it with a new MyTasks component for better organization and readability.
- Created a TaskCard component to encapsulate individual task display logic, including status updates.
- Updated the sidebar configuration with new user profile information.
- Streamlined task management functionality by simulating status updates within the new components.
