+++ frontend/src/lib/sidebar-config.tsx
@@ -122,9 +122,9 @@ export const sidebarConfig: SidebarSection[] = [
 ];
 
 export const currentUser: UserProfile = {
-  name: "Siriwat K.",
-  email: "siriwatk@test.com",
-  avatar: "https://images.unsplash.com/photo-1535713875002-d1d0cf377fde?auto=format&fit=crop&w=286",
+  name: "AISoc IT",
+  email: "ai.soc.it@gmail.com",
+  avatar: "https://unswaisoc.com/static/media/logo.e6e71831c6d3fb4f7f84.webp",
 };
 
 export interface BrandConfig {
